+++ src/models/folders.ts
@@ -101,6 +101,13 @@ export interface ListFolderQueryParams extends ListQueryParams {
    * (Microsoft and EWS only.) Use the ID of a folder to find all child folders it contains.
    */
   parentId?: string;
+
+  /**
+   * (Microsoft only) If true, retrieves folders from a single-level hierarchy only. 
+   * If false, retrieves folders across a multi-level hierarchy.
+   * @default false
+   */
+  singleLevel?: boolean;
 }
 
 export type UpdateFolderRequest = Partial<CreateFolderRequest>;